feat(gui): Sync setting widgets about frees space warning and remove less free space
The values of two widgets in Manage Profile dialog depend on each other and will be synchronized.
In Options tab it is 'Warn on free space' and in Remove & Retention tab it is 'Remove if less than'. The latter always need to be lower than the first.

The current implementation does not respect the storage size unit. Will be fixed in next PR.